Use the value prop for Option if it contains a complex markup
The `<Option>` component generally takes a single string. In most of these
cases, Listbox recursively looks for the leaf node's string content. In the
following case, it will find the string "Apple":

```
<Option>
  <div>Apple</div>
</Option>
```

However, ocasionally one might want an Option to contain multiple HTML elements
to represent a more complex UI:

```
<Option>
  <div>Apple</div>
  <span>|</span>
  <div>On Sale!</div>
</Option>
```

So finding a single text content string won't work. In these cases, Option
provides a value prop. The value prop will be used as the  selected item when
a user selects an option.