Update ComboBox Storybook
Updated event listener so that when a user clicks outside of the current combo
box and into another listbox, we ensure that the current combo box popup is
collapsed.

Only set aria-activedescendant after the Down arrow key is used to start
traversing the associated Listbox options, and to clear aria-activedescendant
by removing the attribute or setting it to “” every time the Left/Right arrow
keys are pressed.

See: https://www.levelaccess.com/differences-aria-1-0-1-1-changes-rolecombobox/